Extract modality, intention, preference, and priority
